[ wip ]: test-coverage improvements
